Um baita de um update

+++ Backend/tcc/src/modules/professor/services/DisciplinaServices/DeleteDisciplinaService.ts
@@ -12,10 +12,15 @@ class DeleteDisciplinaService {
     private disciplinaRepository: IDisciplinaRepository,
   ) {}
 
-  public async execute(disciplina_id: string): Promise<Disciplina> {
+  public async execute(professor_id: string, disciplina_id: string): Promise<Disciplina> {
     const result = await this.disciplinaRepository.findByID(disciplina_id);
+
     if (!result) {
       throw new AppError('Disciplina não encontrada');
+    }
+
+    if (professor_id != result.professor_id) {
+      throw new AppError('Disciplina não pertence a esse professor');
     }
 
     await this.disciplinaRepository.deleted(disciplina_id);
